backend/requirements.txt
django-cors-headers
djangorestframework
django
xhtml2pdf
matplotlib
weasyprint
celery==5.2.7
django-celery-beat==2.2.1
redis==4.5.1
